backend/requirements.txt
django>=5.0
psycopg2-binary
python-dotenv
djangorestframework
djangorestframework-simplejwt
drf-spectacular
Pillow
requests
